Prepare for DB rewrite, implement deps and IPC
Database code is now factored out into its own thing, but also not done
(there's only a few lines to write to get some sort of MVP version that
won't do anything super cool but will probably work basically fine).

Task dependency handling should probably work but the database needs to
work to really be able to test that. So that'll be the next thing to do
in order to basically kind of sort of have a useful program.