Add saving/loading, fix remaining silliness
It WORKS!

Also, something I forgot to mention: several commits ago tcc support was
dropped because tcc won't put out reasonable shared objects, apparently.
Fortunately, the fast compile times are no longer as important since we
now have a usable build system!!!!!

This commit also removes blake2b code which is no longer used, and adds
a test for fpath. That isn't currently hooked up to be run due to
laziness but it was helpful in figuring out how to actually implement
fpath correctly, and can be manually compiled and run in the event that
you want to change that code for whatever reason. Regression testing
yay!