Implement -B and fix some wrong assumptions
I always had it in my head that I wouldn't have to test infiles if the
task is rerunning anyway, but of course, not bringing said infiles up to
date makes the task rerun again later when it would otherwise not need
to. This would be avoided entirely with file watching (to be figured out
later) so this isn't really a big deal.