Work around some NetBSD stupidity
The stupidity of poll() over there was causing build to lock up and
spin in a manner that was very annoying to debug, but at least the man
page eventually clarified that their implementation is wrong and I'd
wasted all this debugging time through no fault of my own. Thanks guys!