Go ahead and ISC-license mostly everything
And public domain-ify the rest. Build scripts are all in the public
domain too since they serve as useful documentation. Along with the
documentation.

I was holding off on licensing because I considered making up some less
verbose licence text but I don't think it matters that much and it's
probably better to use something well-established. So ISC it is.