Fix building on illumos
Finally solved the mystery that was breaking the VM, so was able to test
and get this working.

Amazingly, no code changes (well, except -lsocket, totally forgot that
was a thing). Just had to fix a little bit of weirdness in some scripts.
Oh, and add a hack to libcpoly to fix a warning, although amusingly that
warning only actually happened inside libcpoly itself. Worth fixing
anyway though since mmap is kind of useful in general.

Next up macOS support. Just kidding.